Time: 2 ms (60.92%), Space: 6.1 MB (6.02%) - LeetHub

+++ 342-power-of-four/342-power-of-four.cpp
@@ -3,7 +3,7 @@ public:
     bool isPowerOfFour(int n) {
         if(n == 0)
             return false;
-        float power = log(n)/log(4);
+        double power = log(n)/log(4);
         if(power == floor(power))
             return true;
         else return false;
